Add outer container around the announcements container
Adding the outer container around the announcements
container makes it possible for the "fractional"
bottom border effect that each of the announcement
cards have (except for the last one). The padding
around the whole announcements section is applied to
the outer container such that the announcements section
appears in alignment with the projects section's
content and padding is applied to the announcements
container itself to create the border effect.